JMSeq State Machine implementation
We use Continuent State Machine implementation to build a simple FSM for the call expressions in JMSeq. We issue events based
on call expressions that are built on receving the execution events. This commit will be an important one in the direction of
correct state machine simulation for verifying the executions.